[feat] : RacingCar class에 name 필드 추가
